HLSL signature parsing now happening in a compiler, and cached to intermediates - should hopefully allow us to avoid putting shaders in the end project, and just include the intermediates instead - also shader compilation no longer uses DependentFileStates directly, and instead just tracks dependencies via DepVal - since we can work backwards from the DepVal to the DependentFileStates now

+++ ShaderParser/NodeGraphSignature.cpp
@@ -45,7 +45,7 @@ namespace GraphLanguage
 	static void SerializationOperator(::Assets::BlockSerializer& serializer, const NodeGraphSignature::Parameter& sig)
 	{
 		serializer << sig._type << sig._name;
-		serializer << (uint32_t)sig._direction;
+		serializer << (uint64_t)sig._direction;
 		serializer << sig._semantic << sig._default;
 	}
